Add files via upload
In order to do the plotting points and showing the description, I guess It requires the graph.
Made the graph but it's not done yet. 

Had some troubles while doing it such as:
"WARNING: Loading FXML document with JavaFX API of version 9.0.1 by JavaFX runtime of version 8.0.151"
Had to restart from the beginning couple times.
I put some random number or string to test the graph to show the labels , but somehow it is not working.
Moreover, I couldn't use the database(mysql) due to my lack of knowledge.
So, couldn't plotting the number that generates from task #1 (random generator).

I will keep working on it and modify  it
had found resources for other subtasks like event handler (Showing blood sugar level  & time.)

I guess,  if I figure out linking the database / RandomSugarTimer with the Graph, it will in good shape.